fixed comment related to changes in signin component and addresses disabled check fixed

+++ src/app/main/component/auth/components/sign-in/sign-in.component.ts
@@ -149,10 +149,8 @@ export class SignInComponent implements OnInit, OnDestroy {
 
   private checkUBSuserData() {
     const ubsData = this.localeStorageService.getUbsPersonalData();
-    if (ubsData) {
-      if (ubsData.email !== this.jwtService.getEmailFromAccessToken()) {
-        this.localeStorageService.deleteUbsPersonalData();
-      }
+    if (ubsData && ubsData.email !== this.jwtService.getEmailFromAccessToken()) {
+      this.localeStorageService.deleteUbsPersonalData();
     }
   }
